requirements.txt
matplotlib
certifi==2017.4.17
chardet==3.0.4
dateutils==0.6.6
future==0.16.0
idna==2.5
mongoengine>=0.13.0
olefile==0.44
pylzma==0.4.8
pymongo==3.4.0
python-dateutil
rarfile==3.0
requests==2.20.1
selenium==3.4.3
six==1.10.0
slackclient==1.0.6
urllib3>=1.23
virtualenv==15.1.0
websocket-client==0.44.0
xvfbwrapper==0.2.9
robobrowser>=0.5
pyfolio
theano
pymc3
numpy
prettytable
tabulate
pandas_datareader
texttable
platypus-opt
python-telegram-bot
pillow>=8.1.2
lxml>=4.6.3
yfinance
